[Tech] rename components
